kubernetes/node-init: do not run script on an already setup node
If the DaemonSet is accidentally removed it can be re-added without
fearing the script will be executed again.

However, if the node is restarted, or upgraded, the node-init script
might need to be ran again which node-init will detected based on the
existence of a file stored in the `/tmp` directory.

Signed-off-by: André Martins <andre@cilium.io>